% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab7.R
\docType{package}
\name{lab7}
\alias{lab7}
\alias{lab7-package}
\title{Rcourse Lab 4 lab7.
Package contains implementation of ridgereg Regression with function print(), predict(), coef().
These function are implmentated in ridgereg class which is created using Reference Class(RC)}
\description{
Rcourse Lab 4 lab7.
Package contains implementation of ridgereg Regression with function print(), predict(), coef().
These function are implmentated in ridgereg class which is created using Reference Class(RC)
}
\section{ridgereg}{

data(iris)
ridgereg(formula=Petal.Length ~ Sepal.Width + Sepal.Length,data=iris)
}

\seealso{
\link[lab7]{ridgereg}
}
