% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\docType{class}
\name{ridgereg}
\alias{ridgereg}
\title{Linear Regression}
\description{
RC class mainpluate linear algeric caluclautions giving formula and data.
}
\section{Fields}{

\describe{
\item{\code{formula}}{Formula}

\item{\code{data}}{A data frame}
}}

\section{Methods}{

\describe{
\item{\code{coef()}}{This function returns the vector of beta coefficients}

\item{\code{initialize(formula, data, lambda = 0, normalize = FALSE)}}{This function acts as constructor}

\item{\code{predict()}}{This function returns the vector of calculated fitted values}

\item{\code{print()}}{This function prints the formula and dataset name as well as the calculated coefficients}
}}

