% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statPack.R
\docType{package}
\name{statPack}
\alias{statPack}
\alias{statPack-package}
\title{Rcourse Lab 7 ridereg
Package contains implementation of Linear Regression with function print(), summary(), plot(), resid(), coef().
These function are implmentated in linreg class which is created using Reference Class(RC)}
\description{
Rcourse Lab 7 ridereg
Package contains implementation of Linear Regression with function print(), summary(), plot(), resid(), coef().
These function are implmentated in linreg class which is created using Reference Class(RC)
}
\section{ridgereg}{


ridgereg(formula=Petal.Length ~ Sepal.Width + Sepal.Length,data=iris)
}

\seealso{
\link[statPack]{ridgereg}
}
